refactor: rename bounce-helpers to bounce-test-helpers
